moved team.api to api folder
